Inline router utilities (#641)
